Assert that the flattened PDF has the desired text
This is needed because openpdf (like pdfbox) gives no indication if the
text field cannot render the characters. It always returns the set value
on a subsequent get. Only by flattening the PDF and ensuring that the
text is present can we ensure that the font substitution happened
correctly.

Also removed test that asserts that nulls are set as empty string, as
openpdf retains the null and the net effect is the same.