Automatically indicate required input fields (#502)
Automatically show required fields in form fields. Uses @NotBlank, @NotEmpty and @NotNull annotations on input fields to decide which fields should be marked as required in the client side form.

Adds a mechanism to mark an input as required using an attribute `required` passed to the thymeleaf fragment for the input. 

---------

Co-authored-by: Sree P <sprasad@codeforamerica.org>